+++ packages/meteor-developer-oauth/package.js
@@ -1,12 +1,12 @@
 Package.describe({
   summary: 'Meteor developer accounts OAuth flow',
-  version: '1.3.0'
+  version: '1.3.1'
 });
 
 Package.onUse(api => {
   api.use('oauth2', ['client', 'server']);
   api.use('oauth', ['client', 'server']);
-  api.use('http', ['server']);
+  api.use('http@1.4.4 || 2.0.0', ['server']);
   api.use(['ecmascript', 'service-configuration'], ['client', 'server']);
   api.use('random', 'client');