+++ packages/oauth1/package.js
@@ -10,7 +10,7 @@ Package.onUse(api => {
   api.use('oauth', ['client', 'server']);
   api.use([
     'check',
-    'http@2.0.0'
+    'http@1.4.4 || 2.0.0'
   ], 'server');
 
   api.use('mongo');